Re-arrange the package structure
Moves OutputFormat and AssetBuilder into the asset subpackage. Moves
Implementations into formatters and builders respectively. Makes the
minimal code changes required to for everything to work.
